Updating the DagMC test to conform to API change.

+++ tools/dagmc/ray_fire_test.cc
@@ -500,9 +500,9 @@ void write_obbtree_histogram( EntityHandle root, OrientedBoxTreeTool& tree, std:
 
 }
 
-void moab_memory_estimates( Interface* mbi, int& moab_data_bytes, int& moab_alldata_est_bytes ){
+void moab_memory_estimates( Interface* mbi, unsigned long long& moab_data_bytes, unsigned long long& moab_alldata_est_bytes ){
   
-  unsigned long storage, amortized_storage;
+  unsigned long long storage, amortized_storage;
   mbi->estimated_memory_use( NULL, 0, &storage, &amortized_storage );
 
   moab_data_bytes = storage;
@@ -550,7 +550,7 @@ void dump_pyfile( char* filename, double timewith, double timewithout, double tm
     DICT_VAL(timewith-timewithout);
   }
   DICT_VAL(tmem);
-  int moab_data_bytes, moab_alldata_est_bytes;
+  unsigned long long moab_data_bytes, moab_alldata_est_bytes;
   moab_memory_estimates( dagmc.moab_instance(), moab_data_bytes, moab_alldata_est_bytes );
   DICT_VAL( moab_data_bytes );
   DICT_VAL( moab_alldata_est_bytes );
